requirements.txt
aiocontextvars==0.2.2
aiosqlite==0.17.0
amqp==5.1.1
aniso8601==9.0.1
anyio==3.6.2
arabic-reshaper==2.1.4
argon2-cffi==21.1.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
autopep8==2.0.1
backcall==0.2.0
billiard==3.6.4.0
black==22.6.0
bleach==4.1.0
boto3==1.23.10
botocore==1.26.10
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
contextlib2==21.6.0
contextvars==2.4
coreapi==2.3.3
coreschema==0.0.4
cryptocode==0.1
cycler==0.10.0
dataclasses==0.8
decorator==5.1.0
defusedxml==0.7.1
Django==3.2
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
entrypoints==0.3
fastapi==0.83.0
future==0.18.2
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.0
gunicorn==20.1.0
h11==0.13.0
html5lib==1.1
idna==3.3
immutables==0.16
importlib-metadata==4.8.1
inflection==0.5.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso8601==1.0.2
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.19.5
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.21
psycopg2==2.7.5
psycopg2-binary==2.8
pycodestyle==2.10.0
pycparser==2.20
pycryptodomex==3.16.0
pydantic==1.9.2
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==3.0.0
PyPDF2==2.12.1
PyPika==0.48.8
pyrsistent==0.18.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
pywin32==302
pywinpty==1.1.4
pyzmq==22.3.0
qrcode==7.3.1
qtconsole==5.1.1
QtPy==1.11.2
redis==4.3.5
reportlab==3.6.6
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
starlette==0.19.1
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
tomli==1.2.3
tornado==6.1
tortoise-orm==0.15.24
traitlets==4.3.3
typed-ast==1.5.4
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.16.0
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
windows-curses==2.3.1
xhtml2pdf==0.2.5
xlwt==1.3.0
xmltodict==0.13.0
zipp==3.6.0
